+++ server/src/test/java/org/opensearch/index/query/QueryStringQueryBuilderTests.java
@@ -34,6 +34,7 @@ package org.opensearch.index.query;
 
 import org.apache.lucene.document.LongPoint;
 import org.apache.lucene.document.SortedNumericDocValuesField;
+import org.apache.lucene.document.SortedSetDocValuesField;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.queries.BlendedTermQuery;
 import org.apache.lucene.queries.spans.SpanNearQuery;
@@ -502,7 +503,13 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
     public void testToQueryMultipleFieldsBooleanQuery() throws Exception {
         Query query = queryStringQuery("test").field(TEXT_FIELD_NAME).field(KEYWORD_FIELD_NAME).toQuery(createShardContext());
         Query expected = new DisjunctionMaxQuery(
-            List.of(new TermQuery(new Term(TEXT_FIELD_NAME, "test")), new TermQuery(new Term(KEYWORD_FIELD_NAME, "test"))),
+            List.of(
+                new TermQuery(new Term(TEXT_FIELD_NAME, "test")),
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "test")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("test"))
+                )
+            ),
             0
         );
         assertEquals(expected, query);
@@ -511,7 +518,13 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
     public void testToQueryMultipleFieldsDisMaxQuery() throws Exception {
         Query query = queryStringQuery("test").field(TEXT_FIELD_NAME).field(KEYWORD_FIELD_NAME).toQuery(createShardContext());
         Query expected = new DisjunctionMaxQuery(
-            List.of(new TermQuery(new Term(TEXT_FIELD_NAME, "test")), new TermQuery(new Term(KEYWORD_FIELD_NAME, "test"))),
+            List.of(
+                new TermQuery(new Term(TEXT_FIELD_NAME, "test")),
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "test")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("test"))
+                )
+            ),
             0
         );
         assertEquals(expected, query);
@@ -520,7 +533,13 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
     public void testToQueryFieldsWildcard() throws Exception {
         Query query = queryStringQuery("test").field("mapped_str*").toQuery(createShardContext());
         Query expected = new DisjunctionMaxQuery(
-            List.of(new TermQuery(new Term(TEXT_FIELD_NAME, "test")), new TermQuery(new Term(KEYWORD_FIELD_NAME, "test"))),
+            List.of(
+                new TermQuery(new Term(TEXT_FIELD_NAME, "test")),
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "test")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("test"))
+                )
+            ),
             0
         );
         assertEquals(expected, query);
@@ -545,7 +564,10 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
         Query expected = new DisjunctionMaxQuery(
             List.of(
                 new BoostQuery(new TermQuery(new Term(TEXT_FIELD_NAME, "test")), 2.2f),
-                new TermQuery(new Term(KEYWORD_FIELD_NAME, "test"))
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "test")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("test"))
+                )
             ),
             0
         );
@@ -1006,7 +1028,12 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
             ).add(new BooleanClause(new TermQuery(new Term(TEXT_FIELD_NAME, "bar")), BooleanClause.Occur.SHOULD)).build();
             List<Query> disjuncts = new ArrayList<>();
             disjuncts.add(bq1);
-            disjuncts.add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo bar")));
+            disjuncts.add(
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo bar")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("foo bar"))
+                )
+            );
             DisjunctionMaxQuery expectedQuery = new DisjunctionMaxQuery(disjuncts, 0.0f);
             assertThat(query, equalTo(expectedQuery));
         }
@@ -1020,7 +1047,12 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
             List<Query> disjuncts = new ArrayList<>();
             PhraseQuery pq = new PhraseQuery.Builder().add(new Term(TEXT_FIELD_NAME, "foo")).add(new Term(TEXT_FIELD_NAME, "bar")).build();
             disjuncts.add(pq);
-            disjuncts.add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo bar")));
+            disjuncts.add(
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo bar")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("foo bar"))
+                )
+            );
             DisjunctionMaxQuery expectedQuery = new DisjunctionMaxQuery(disjuncts, 0.0f);
             assertThat(query, equalTo(expectedQuery));
         }
@@ -1034,7 +1066,12 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
             ).add(new BooleanClause(new TermQuery(new Term(TEXT_FIELD_NAME, "bar")), BooleanClause.Occur.SHOULD)).build();
             List<Query> disjuncts = new ArrayList<>();
             disjuncts.add(bq1);
-            disjuncts.add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo bar")));
+            disjuncts.add(
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo bar")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("foo bar"))
+                )
+            );
             DisjunctionMaxQuery disjunctionMaxQuery = new DisjunctionMaxQuery(disjuncts, 0.0f);
             BooleanQuery expectedQuery = new BooleanQuery.Builder().add(disjunctionMaxQuery, BooleanClause.Occur.SHOULD)
                 .add(new TermQuery(new Term(TEXT_FIELD_NAME, "other")), BooleanClause.Occur.SHOULD)
@@ -1049,12 +1086,22 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
 
             List<Query> disjuncts1 = new ArrayList<>();
             disjuncts1.add(new TermQuery(new Term(TEXT_FIELD_NAME, "foo")));
-            disjuncts1.add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo")));
+            disjuncts1.add(
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("foo"))
+                )
+            );
             DisjunctionMaxQuery maxQuery1 = new DisjunctionMaxQuery(disjuncts1, 0.0f);
 
             List<Query> disjuncts2 = new ArrayList<>();
             disjuncts2.add(new TermQuery(new Term(TEXT_FIELD_NAME, "bar")));
-            disjuncts2.add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "bar")));
+            disjuncts2.add(
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "bar")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("bar"))
+                )
+            );
             DisjunctionMaxQuery maxQuery2 = new DisjunctionMaxQuery(disjuncts2, 0.0f);
 
             BooleanQuery expectedQuery = new BooleanQuery.Builder().add(new BooleanClause(maxQuery1, BooleanClause.Occur.SHOULD))
@@ -1295,7 +1342,13 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
             Query expected = new DisjunctionMaxQuery(
                 Arrays.asList(
                     new TermQuery(new Term(TEXT_FIELD_NAME, "hello")),
-                    new BoostQuery(new TermQuery(new Term(KEYWORD_FIELD_NAME, "hello")), 5.0f)
+                    new BoostQuery(
+                        new IndexOrDocValuesQuery(
+                            new TermQuery(new Term(KEYWORD_FIELD_NAME, "hello")),
+                            SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("hello"))
+                        ),
+                        5.0f
+                    )
                 ),
                 0.0f
             );
@@ -1351,7 +1404,10 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
             new QueryStringQueryBuilder("bar").quoteFieldSuffix("_2").field(TEXT_FIELD_NAME).doToQuery(context)
         );
         assertEquals(
-            new TermQuery(new Term(KEYWORD_FIELD_NAME, "bar")),
+            new IndexOrDocValuesQuery(
+                new TermQuery(new Term(KEYWORD_FIELD_NAME, "bar")),
+                SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("bar"))
+            ),
             new QueryStringQueryBuilder("\"bar\"").quoteFieldSuffix("_2").field(TEXT_FIELD_NAME).doToQuery(context)
         );
 
@@ -1392,8 +1448,20 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
                 new BooleanQuery.Builder().add(new TermQuery(new Term(TEXT_FIELD_NAME, "quick")), Occur.SHOULD)
                     .add(new TermQuery(new Term(TEXT_FIELD_NAME, "fox")), Occur.SHOULD)
                     .build(),
-                new BooleanQuery.Builder().add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "quick")), Occur.SHOULD)
-                    .add(new TermQuery(new Term(KEYWORD_FIELD_NAME, "fox")), Occur.SHOULD)
+                new BooleanQuery.Builder().add(
+                    new IndexOrDocValuesQuery(
+                        new TermQuery(new Term(KEYWORD_FIELD_NAME, "quick")),
+                        SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("quick"))
+                    ),
+                    Occur.SHOULD
+                )
+                    .add(
+                        new IndexOrDocValuesQuery(
+                            new TermQuery(new Term(KEYWORD_FIELD_NAME, "fox")),
+                            SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("fox"))
+                        ),
+                        Occur.SHOULD
+                    )
                     .build()
             ),
             0f
@@ -1451,22 +1519,39 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
                 )
             );
         try {
-            Term[] blendedTerms = new Term[2];
-            blendedTerms[0] = new Term(TEXT_FIELD_NAME, "foo");
-            blendedTerms[1] = new Term(KEYWORD_FIELD_NAME, "foo");
+            Term[] terms = new Term[1];
+            terms[0] = new Term(TEXT_FIELD_NAME, "foo");
 
             Query query = new QueryStringQueryBuilder("foo").analyzer("whitespace")
                 .type(MultiMatchQueryBuilder.Type.CROSS_FIELDS)
                 .toQuery(createShardContext());
-            Query expected = BlendedTermQuery.dismaxBlendedQuery(blendedTerms, 1.0f);
+            List<Query> queries = new ArrayList<>(terms.length);
+            Query textFieldExpected = BlendedTermQuery.dismaxBlendedQuery(terms, 1.0f);
+            Query keywordFieldExpected = new IndexOrDocValuesQuery(
+                new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo")),
+                SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("foo"))
+            );
+            queries.add(textFieldExpected);
+            queries.add(keywordFieldExpected);
+            Query expected = new DisjunctionMaxQuery(queries, 0);
             assertEquals(expected, query);
 
             query = new QueryStringQueryBuilder("foo mapped_string:10").analyzer("whitespace")
                 .type(MultiMatchQueryBuilder.Type.CROSS_FIELDS)
                 .toQuery(createShardContext());
-            expected = new BooleanQuery.Builder().add(BlendedTermQuery.dismaxBlendedQuery(blendedTerms, 1.0f), Occur.SHOULD)
-                .add(new TermQuery(new Term(TEXT_FIELD_NAME, "10")), Occur.SHOULD)
-                .build();
+            expected = new BooleanQuery.Builder().add(
+                new DisjunctionMaxQuery(
+                    List.of(
+                        new IndexOrDocValuesQuery(
+                            new TermQuery(new Term(KEYWORD_FIELD_NAME, "foo")),
+                            SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("foo"))
+                        ),
+                        BlendedTermQuery.dismaxBlendedQuery(terms, 1.0f)
+                    ),
+                    0
+                ),
+                Occur.SHOULD
+            ).add(new TermQuery(new Term(TEXT_FIELD_NAME, "10")), Occur.SHOULD).build();
             assertEquals(expected, query);
         } finally {
             // Reset the default value
@@ -1535,7 +1620,15 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
             .toQuery(createShardContext());
         List<Query> terms = new ArrayList<>();
         terms.add(new BoostQuery(new TermQuery(new Term(TEXT_FIELD_NAME, "first")), 0.075f));
-        terms.add(new BoostQuery(new TermQuery(new Term(KEYWORD_FIELD_NAME, "first")), 0.5f));
+        terms.add(
+            new BoostQuery(
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "first")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("first"))
+                ),
+                0.5f
+            )
+        );
         Query expected = new DisjunctionMaxQuery(terms, 1.0f);
         assertEquals(expected, query);
     }
@@ -1557,7 +1650,13 @@ public class QueryStringQueryBuilderTests extends AbstractQueryTestCase<QueryStr
         assertEquals(9, noMatchNoDocsQueries);
         assertThat(
             disjunctionMaxQuery.getDisjuncts(),
-            hasItems(new TermQuery(new Term(TEXT_FIELD_NAME, "hello")), new TermQuery(new Term(KEYWORD_FIELD_NAME, "hello")))
+            hasItems(
+                new TermQuery(new Term(TEXT_FIELD_NAME, "hello")),
+                new IndexOrDocValuesQuery(
+                    new TermQuery(new Term(KEYWORD_FIELD_NAME, "hello")),
+                    SortedSetDocValuesField.newSlowExactQuery(KEYWORD_FIELD_NAME, new BytesRef("hello"))
+                )
+            )
         );
     }