+++ server/src/test/java/org/opensearch/search/sort/FieldSortBuilderTests.java
@@ -40,11 +40,13 @@ import org.apache.lucene.document.FloatPoint;
 import org.apache.lucene.document.IntPoint;
 import org.apache.lucene.document.LongPoint;
 import org.apache.lucene.document.SortedNumericDocValuesField;
+import org.apache.lucene.document.SortedSetDocValuesField;
 import org.apache.lucene.document.TextField;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.sandbox.document.BigIntegerPoint;
 import org.apache.lucene.sandbox.document.HalfFloatPoint;
+import org.apache.lucene.search.IndexOrDocValuesQuery;
 import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.SortField;
 import org.apache.lucene.search.SortedNumericSelector;
@@ -319,7 +321,13 @@ public class FieldSortBuilderTests extends AbstractSortTestCase<FieldSortBuilder
         XFieldComparatorSource comparatorSource = (XFieldComparatorSource) sortField.getComparatorSource();
         Nested nested = comparatorSource.nested();
         assertNotNull(nested);
-        assertEquals(new TermQuery(new Term(MAPPED_STRING_FIELDNAME, "value")), nested.getInnerQuery());
+        assertEquals(
+            new IndexOrDocValuesQuery(
+                new TermQuery(new Term(MAPPED_STRING_FIELDNAME, "value")),
+                SortedSetDocValuesField.newSlowExactQuery(MAPPED_STRING_FIELDNAME, new BytesRef("value"))
+            ),
+            nested.getInnerQuery()
+        );
 
         sortBuilder = new FieldSortBuilder("fieldName").setNestedPath("path");
         sortField = sortBuilder.build(shardContextMock).field;
@@ -336,7 +344,13 @@ public class FieldSortBuilderTests extends AbstractSortTestCase<FieldSortBuilder
         comparatorSource = (XFieldComparatorSource) sortField.getComparatorSource();
         nested = comparatorSource.nested();
         assertNotNull(nested);
-        assertEquals(new TermQuery(new Term(MAPPED_STRING_FIELDNAME, "value")), nested.getInnerQuery());
+        assertEquals(
+            new IndexOrDocValuesQuery(
+                new TermQuery(new Term(MAPPED_STRING_FIELDNAME, "value")),
+                SortedSetDocValuesField.newSlowExactQuery(MAPPED_STRING_FIELDNAME, new BytesRef("value"))
+            ),
+            nested.getInnerQuery()
+        );
 
         // if nested path is missing, we omit any filter and return a SortedNumericSortField
         sortBuilder = new FieldSortBuilder("fieldName").setNestedFilter(QueryBuilders.termQuery(MAPPED_STRING_FIELDNAME, "value"));